move validators and types inside each directory componenet
