+++ src/components/LcIcon/LcIcon.stories.ts
@@ -1,4 +1,4 @@
-import { vSize } from '../../validators/LcIcon'
+import { vSize } from './validators/LcIcon'
 import LcIcon from './LcIcon'
 
 export default {
