+++ src/components/LcIcon/validators/LcIcon.ts
