+++ src/components/LcTooltip/validators/LcTooltip.ts
